Move LogWrapper out of internal package
