+++ libs/cardreader/src/main/java/com/woocommerce/android/cardreader/LogWrapper.kt
@@ -1,6 +1,6 @@
-package com.woocommerce.android.cardreader.internal.wrappers
+package com.woocommerce.android.cardreader
 
-internal interface LogWrapper {
+interface LogWrapper {
     fun w(tag: String, message: String)
     fun d(tag: String, message: String)
     fun e(tag: String, message: String)
